backend/simulator-api/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.30.6
requests==2.31.0
python-dotenv==1.0.1
jinja2==3.1.2
streamlit==1.41.0
pandas==2.2.3
matplotlib==3.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4